api/pyproject.toml
[project]
name = "tasks api"
version = "0.0.1"
description = "basic test task"
authors = [
    {name = "urm8", email = "arctgisnil@gmail.com"},
]
dependencies = [
    "django>=4.0.3",
    "celery>=5.2.3",
    "django-ninja>=0.17.0",
    "dj-database-url>=0.5.0",
    "factory-boy>=3.2.1",
    "pytest-django>=4.5.2",
    "gunicorn>=20.1.0",
    "furl>=2.1.3",
    "psycopg2>=2.9.3",
    "requests>=2.27.1",
    "gevent>=21.12.0",
    "redis>=4.1.4",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[project.urls]
Homepage = ""
[tool.pdm]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"


[tool.black]
addopts = '--fast'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | __pypackages__
  )/
  | manage.py           # also separately exclude a file named foo.py in
)
'''
line-length = 119
target_version = ['py310']

[tool.isort]
known_first_party = ["tasks_api", "tasks"]
known_third_party = ["django", "celery", "ninja", "dj_database_url"]
line_length = 119
profile = "black"
skip = "migrations,templates"
skip_gitignore = "True"
force_single_line = "True"

[tool.pylint.master]
ignore = "tests.py, urls.py, migrations, __pycache__, certs test*.py"
load-plugins = "pylint_django"
max-line-length = "119"

[tool.pylint.message_control]
disable = '''
    bad-continuation,
    import-outside-toplevel,
    invalid-name,
    line-too-long,
    logging-format-interpolation,
    missing-class-docstring,
    missing-function-docstring,
    missing-module-docstring,
    too-few-public-methods,
    wrong-import-position,
    ungrouped-imports,
    wrong-import-order
'''
enable = "unused-import, unused-variable, duplicate-code"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tasks_api.settings"
addopts = "--disable-warnings -vvv --durations=25 --no-migrations --reuse-db"
python_files = "test_*.py"
test_paths = "tests"
norecursedirs = "__pypackages__"

[tool.vulture]
exclude = ["**/migrations/*.py", "factory.py"]
make_whitelist = true
min_confidence = 80
sort_by_size = true
verbose = true
